tracking/requirements.txt
bytecode==0.15.1
easydict==1.10
jpeg4py==0.1.4
lmdb==1.4.1
matplotlib==3.7.2
numba==0.58.1
numpy==1.23.5
opencv_python==4.8.0.74
pandas==2.0.3
Pillow==9.4.0
Pillow==10.3.0
pycocotools==2.0.6
PyYAML==6.0.1
PyYAML==6.0.1
six==1.16.0
tensorboardX==2.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
tikzplotlib==0.10.1
timm==0.9.5
torch==2.0.0
torchvision==0.15.0
tqdm==4.65.0
visdom==0.2.4
vot==1
wandb==0.16.5
